Move ConfigurationError to top level
This error is used by providers, not just the dataflow
code where it previously lived.
